Misc code cleanup
Resolved an issue with uninitialized struct members
Removed some redundant conditional checks
Made the boolean part of JS function RegisterSkill( <skillnum>, <bool> ) actually work; can now be used to disable/enable JS-implemented skills
Fixed a bug with cHTMLTemplate::Process() where the same expression was used for both branches of a ternary operator
Various minor cleanups; removed some unused variables, addressed some potential portability issues

+++ source/cmdtable.cpp
@@ -781,7 +781,7 @@ void command_spawnkill( CSocket *s )
 
 void BuildWhoGump( CSocket *s, UI08 commandLevel, std::string title )
 {
-	size_t j = 0;
+	unsigned int j = 0;
 	char temp[512];
 
 	GumpDisplay Who( s, 400, 300 );
